web_application/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
graphqlclient==0.2.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
obs-websocket-py==0.5.1
py-slippi==1.4.0
py-ubjson==0.14.0
requests==2.23.0
six==1.14.0
termcolor==1.1.0
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0